vault: retry support, better error messages
