+++ vault/pyproject.toml
@@ -16,6 +16,7 @@ python = "^3.9"
 cryptography = "^41.0.1"
 requests = "^2.31"
 pyyaml = "^6.0"
+tenacity = "^8.0.1"
 
 [tool.poetry.group.dev.dependencies]
 robocorp-devutils = {path = "../devutils/", develop = true}
