+++ vault/src/robocorp/vault/__init__.py
@@ -37,7 +37,6 @@ Example content of local secrets file:
 }
 ```
 
-
 OR
 
 ```yaml
